# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/facebook/callback": {
            "get": {
                "description": "Handles the callback from Facebook after the user has authenticated.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Facebook OAuth callback",
                "responses": {
                    "200": {
                        "description": "JWT token for the authenticated user",
                        "schema": {
                            "$ref": "#/definitions/API.JWTResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/facebook/logout": {
            "get": {
                "description": "Logs the user out of the Facebook session.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Facebook OAuth logout",
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "schema": {
                            "$ref": "#/definitions/API.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/google/callback": {
            "get": {
                "description": "Handles the callback from Google after the user has authenticated.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Google OAuth callback",
                "responses": {
                    "200": {
                        "description": "JWT token for the authenticated user",
                        "schema": {
                            "$ref": "#/definitions/API.JWTResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/google/logout": {
            "get": {
                "description": "Logs the user out of the Google session.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Google OAuth logout",
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "schema": {
                            "$ref": "#/definitions/API.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "API.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "API.JWTResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "API.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}